journalistic-entity-extraction/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
py2neo
python-dotenv
